requirements.txt
python-telegram-bot==13.12
Pillow==8.3.1
requests==2.26.0
